+++ sdk/src/org.graalvm.polyglot/src/org/graalvm/polyglot/Engine.java
@@ -68,6 +68,7 @@ import java.util.function.Predicate;
 import java.util.logging.Handler;
 import java.util.logging.Level;
 
+import org.graalvm.collections.UnmodifiableEconomicSet;
 import org.graalvm.options.OptionDescriptor;
 import org.graalvm.options.OptionDescriptors;
 import org.graalvm.polyglot.PolyglotException.StackFrame;
@@ -634,6 +635,11 @@ public final class Engine implements AutoCloseable {
         @Override
         public void setHostAccessImpl(HostAccess conf, Object impl) {
             conf.impl = impl;
+        }
+
+        @Override
+        public UnmodifiableEconomicSet<String> getAccessibleLanguages(PolyglotAccess access, String language) {
+            return access.getAccessibleLanguages(language);
         }
 
     }