+++ truffle/src/com.oracle.truffle.api.test/src/com/oracle/truffle/api/test/polyglot/ContextPreInitializationTest.java
@@ -1199,7 +1199,7 @@ public class ContextPreInitializationTest {
 
         protected void useLanguage(CountingContext context, String id) {
             com.oracle.truffle.api.source.Source source = com.oracle.truffle.api.source.Source.newBuilder(id, "", "").build();
-            context.environment().parse(source);
+            context.environment().parsePublic(source);
         }
 
         @CompilerDirectives.TruffleBoundary