+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/impl/Accessor.java
@@ -219,7 +219,7 @@ public abstract class Accessor {
 
         public abstract Object getCurrentVM();
 
-        public abstract CallTarget parseForLanguage(Object vmObject, Source source, String[] argumentNames);
+        public abstract CallTarget parseForLanguage(Object vmObject, Source source, String[] argumentNames, boolean allowInternal);
 
         public abstract Env getEnvForInstrument(Object vm, String languageId, String mimeType);